Add minHeight to the selector
Add a `minHeight` to the selector when `size: <number>` is being used in a blueprint file. By doing so the text ‘No matching files yet.’ will be visible, without a `minHeight` the whole box collapes and only its borders are visible. The value here is hardcoded, it’s probably better to caculate it based on the `.selector-item-empty` class plus borders.